requirements.txt
Flask
Pillow
gunicorn
supabase
app = Flask(__name__)